move types inside substrate dir
