+++ packages/sdk/src/chains/Substrate/__test__/deposit.test.ts
@@ -2,8 +2,8 @@ import { ApiPromise } from '@polkadot/api';
 import { BN } from '@polkadot/util';
 import { web3FromAddress } from '@polkadot/extension-dapp';
 import { InjectedAccountWithMeta } from '@polkadot/extension-inject/types';
-import { XcmMultiAssetIdType } from '../../../types';
-import {DepositCallbacksType} from '../utils/depositFns'
+import { XcmMultiAssetIdType } from '../types';
+import { DepositCallbacksType } from '../utils/depositFns';
 
 import * as Utils from '../utils/depositFns';
