+++ packages/sdk/src/chains/Substrate/index.ts
@@ -1,4 +1,10 @@
 export {
+  XcmMultiAssetIdType,
+  SubstrateConfigAssetType,
+  SubstrateConfigType,
+} from '../Substrate/types';
+
+export {
   substrateSocketConnect,
   getAssetBalance,
   getBasicFee,
